[Engine][Task] fix ci problem
