+++ seatunnel-engine/seatunnel-engine-client/src/test/java/org/apache/seatunnel/engine/client/SeaTunnelClientTest.java
@@ -79,10 +79,10 @@ public class SeaTunnelClientTest {
         JobProxy jobProxy;
         try {
             jobProxy = jobExecutionEnv.execute();
+            Assert.assertNotNull(jobProxy);
         } catch (ExecutionException | InterruptedException e) {
             throw new RuntimeException(e);
         }
-        Assert.assertNotNull(jobProxy);
     }
 
     @After
