+++ seatunnel-engine/seatunnel-engine-client/src/test/resources/fakesource_to_file_complex.conf
@@ -31,7 +31,13 @@ source {
     FakeSource {
       result_table_name = "fake"
       field_name = "name,age",
-      parallelism = 1
+      parallelism = 3
+    }
+
+    FakeSource {
+      result_table_name = "fake"
+      field_name = "name,age",
+      parallelism = 3
     }
 
   # If you would like to get more information about how to configure seatunnel and see full list of source plugins,
